requirements.txt
sh
path.py

# testing
pyfakefs
mock 
nose
